cleanup constructor strategy. relaxed comparison operators
This is based completely on Tyler's findings/code and our resulting
discussions. We are disabling most conversions, and relying completely
on an extraction-type constructor, with a simplifying overload for the
entire vector case.

more about exactly which constructions, copies, and moves are allowed
will be fleshed out and documented soon.

The comparison operators are allowed to compare by value different
memory type vectors.