Matlab assuming responsibility for batch component data.
Note that we need to talk about why the continuity6 case is being
handled differently than the other PDEs in this component. For some
reason continuity6 required the "old" version of coeff_mat and
continuity6_old, whereas the other PDEs did not. Perhaps hidden in here
is the reason the continuity6 testing proved such a nusience the other
week.