Fixed spreadsheet errors
Spreadsheet would be sized improperly when the minimum level tested was greater than 1. Instead of being sized based on the length of the list of levels to test, it is based on the value of the highest level. This guarantees a spreadsheet large enough to contain the data from all profiled levels.