Add mass matrix generation to tests, fix matlab adaptivity plotting
Matlab adaptivity plotting error was due to analytic solution matrix
not being resized properly, which caused issue in batch chaining during
the wavelet to realspace transform.
Also adjusts default levels for the diffusion1 PDE, and sets the mass
matrix in the PDE class to an initial value (eye of DOF size).